% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adcf-package.R
\docType{package}
\name{adcf-package}
\alias{adcf-package}
\alias{_PACKAGE}
\alias{{adcf}-package}
\title{adcf: Auto distance covariance/correlation}
\description{
Auto distance covariance/correlation tests for dependence among univariate observations.
}
\references{
Phyllis Wan, Richard A. Davis, Goodness-of-fit testing for time
series models via distance covariance, Journal of Econometrics,
Volume 227, Issue 1, 2022, Pages 4-24, ISSN 0304-4076,
https://doi.org/10.1016/j.jeconom.2020.05.008.

Richard A. Davis, Muneya Matsui, Thomas Mikosch, Phyllis Wan
"Applications of distance correlation to time series," Bernoulli,
Bernoulli 24(4A), 3087-3116, (November 2018).

Xiaoming Huo & Gábor J. Székely (2016) Fast Computing for
Distance Covariance, Technometrics, 58:4, 435-447,
DOI: 10.1080/00401706.2015.1054435.

Gábor J. Székely, Maria L. Rizzo, Nail K. Bakirov
"Measuring and testing dependence by correlation of distances," The
Annals of Statistics, Ann. Statist. 35(6), 2769-2794, (December 2007).
}
\author{
\strong{Maintainer}: Leon Fernandes \email{leon.fernandes@columbia.com} (\href{https://orcid.org/0000-0002-7192-9939}{ORCID})

}
\keyword{internal}
