% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adcf.R
\name{adcf}
\alias{adcf}
\title{Auto distance covariance}
\usage{
adcf(x, lags = NULL, mu = c("szekely", "gaussian"), ...)
}
\arguments{
\item{x}{numeric vector.}

\item{lags}{integer. Specifies lags at which distance covariance is to be
calculated. When \code{NULL} maximum range of lags are used.}

\item{mu}{Single character string corresponding to the weight measure to be
used. Must be one of "szekely" or "gaussian".}

\item{...}{Optional parameters supplied for the weight measure \code{mu}.}
}
\value{
A \code{tibble} with columns \code{lag}, \code{auto_dist_covariance} and
\code{auto_dist_correlation}.
}
\description{
Calculates the auto-distance covariance/correlation of \code{x}.
}
\examples{
# adcf ----------------------------------------------------------------------
x <- rnorm(100)
# Adcf values are small for all lags due to independence
adcf(x)

y <- stats::arima.sim(model = list("ar" = 0.6), n = 100)
# Adcf values are larger for initial lags.
adcf(y)
}
