% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adcf_metric.R
\name{adcf_metric}
\alias{adcf_metric}
\alias{adcf_metric.data.frame}
\alias{adcf_metric_vec}
\title{Measures independence of observations based on auto distance covariance}
\usage{
adcf_metric(data, ...)

\method{adcf_metric}{data.frame}(
  data,
  truth,
  ...,
  lag = 1:vctrs::vec_size(data) - 1,
  na_rm = TRUE,
  case_weights = NULL
)

adcf_metric_vec(
  truth,
  estimate,
  lag = 0:1,
  na_rm = TRUE,
  case_weights = NULL,
  ...
)
}
\arguments{
\item{truth}{The column identifier for the true results (that is \code{numeric}).}

\item{lag}{Vector of positive integers. This corresponds to the lag at
which distance covariance should be evaluated.}

\item{estimate}{The column identifier for the predicted resutls (that is also
\code{numeric})}
}
\value{
A \code{tibble} with columns \code{.metric}, \code{.estimator} and \code{.estimate} and
1 row of values. For \code{adcf_metric_vec}, a single \code{numeric} value (or \code{NA}).
}
\description{
Measures independence of observations based on auto distance covariance
}
