% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adcv_to_adcf.R
\name{adcv_to_adcf}
\alias{adcv_to_adcf}
\title{Convert auto distance covariance to auto distance correlation}
\usage{
adcv_to_adcf(object)
}
\arguments{
\item{object}{a tibble of class \code{adcv_tbl} that includes row with auto
ditance covariance at lag 0.}
}
\value{
tibble of class \code{adcf_tbl}.
}
\description{
Convert auto distance covariance to auto distance correlation
}
