% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adcf.R
\name{adcf}
\alias{adcf}
\title{Calculates the auto-distance covariance of \code{x}. The weight measure for the
distance covariance is specified via the Fourier transform function \code{mu_hat}.}
\usage{
adcf(x, lag = NULL, mu = c("szekely", "gaussian"), ...)
}
\arguments{
\item{x}{numeric vector.}

\item{lag}{integer. Specifies lag(s) at which distance covariance is to be
calculated. When \code{NULL} maximum range of lags are used.}

\item{mu}{Single character string corresponding to the weight measure to be
used. Must be one of "szekely" or "gaussian".}

\item{...}{Optional parameters supplied for the weight measure \code{mu}.}
}
\value{
An object of class \code{adcf_tbl}.
}
\description{
Calculates the auto-distance covariance of \code{x}. The weight measure for the
distance covariance is specified via the Fourier transform function \code{mu_hat}.
}
\examples{
# adcf ----------------------------------------------------------------------
x <- rnorm(100)
# Adcf values are small for all lags due to independence
adcf(x)

y <- stats::arima.sim(model = list("ar" = 0.6), n = 100)
# Adcf values are larger for initial lags.
adcf(y)
}
