% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcf.R
\name{dcf}
\alias{dcf}
\title{Calculate pairwise distance covariances}
\usage{
dcf(x, y, lag = c(0L, 1L), mu = c("szekely", "gaussian"), ...)
}
\arguments{
\item{x}{numeric vector.}

\item{y}{numeric vector.}

\item{lag}{integer. Lag to be applied to \code{y}.}

\item{mu}{Single character string corresponding to the weight measure to be
used. Must be one of "szekely" or "gaussian".}

\item{...}{Optional parameters supplied for the weight measure \code{mu}.}
}
\value{
A tibble of lagged distance covariances.
}
\description{
Flexible version of distance covariance which uses the maximum number of
observations from \code{x} and \code{y} for each lag.
}
