% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcf_helpers.R
\name{mu_hat_szekely}
\alias{mu_hat_szekely}
\alias{mu_hat_gaussian}
\alias{get_weight_measure}
\title{Fourier transforms from pairwise distance matrix.}
\usage{
mu_hat_szekely(x, index = 1)

mu_hat_gaussian(x, sigma2 = 0.25)

get_weight_measure(mu = c("szekely", "gaussian"), index = 1, sigma2 = 0.25)
}
\arguments{
\item{x}{Matrix of pairwise distances.}

\item{index}{Numeric between 0 and 2.}

\item{sigma2}{Variance of guassian measure.}

\item{mu}{Single character string of weight measure to be used.
Returns function which Fourier transform according to a given weight measure}
}
\value{
numeric matrix of Fourier transformed version of \code{x} according to
distance covariance weight measure of \cite{Szekelyetal2007}.

numeric matrix of Fourier transformed version of \code{x} according to
Gaussian weight measure.
}
\description{
Fourier transforms from pairwise distance matrix.
}
